Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,7 +5,7 @@ int x,y,z;
 float a;
 scanf("%d %d %d",&x,&y,&z);
 
-a=x+y+z/3;
+a=(float)(x+y+z)/3;
 printf("Average: %f",a);
 return 0;
